+++ src/gui/include/TileRenderer.tcc
@@ -58,9 +58,9 @@ void TileRenderer::RenderAtoms(Drawing & drawing, SPoint& pt, Tile<CC>& tile,
           }
           else
           {
-            for(s32 x = 0; x < m_atomDrawSize; x++)
+            for(s32 x = 0; x < (s32) m_atomDrawSize; x++)
             {
-              for(s32 y = 0; y < m_atomDrawSize; y++)
+              for(s32 y = 0; y < (s32) m_atomDrawSize; y++)
               {
                 if((x + y + 1) & 4)
                 {
